Rename store.js to configureStore.js to Prevent conflict with storeJS npm package
