+++ internals/templates/app.js
@@ -37,7 +37,7 @@ import '!file-loader?name=[name].[ext]!./manifest.json';
 import 'file-loader?name=[name].[ext]!./.htaccess';
 /* eslint-enable import/no-unresolved, import/extensions */
 
-import configureStore from './store';
+import configureStore from './configureStore';
 
 // Import i18n messages
 import { translationMessages } from './i18n';
