+++ internals/templates/configureStore.js
