+++ internals/templates/utils/tests/sagaInjectors.test.js
@@ -5,7 +5,7 @@
 import { memoryHistory } from 'react-router-dom';
 import { put } from 'redux-saga/effects';
 
-import configureStore from '../../store';
+import configureStore from '../../configureStore';
 import getInjectors, {
   injectSagaFactory,
   ejectSagaFactory,
